+++ CHANGELOG.md
@@ -7,3 +7,4 @@ Changelog can generally be found here https://www.spigotmc.org/resources/17067/u
 - Separate /discord command permissions (discordmc.command.#, discordmc.command.login, discordmc.command.logout, discordmc.command.lookup)
 - More reliable connection detection
 - Fix issue with some chat extension plugins
+- Add update notification using Spiget, requires "discordmc.admin"